pyproject.toml
[project]
name = "hetzner-dns-updater"
version = "0.1.0"
description = "DNS updater for hetzner console"
requires-python = ">=3.14"
dependencies = [
    "black>=26.1.0",
    "python-dotenv>=1.2.1",
    "requests==2.32.5",
]
